refactor(cwt): move worktree parsing into its own module
Pure move of Worktree, parse_worktree_list, get_worktrees and paths_equal
out of main.rs so the nested-repository discovery has a home. No behavior
change.
