+++ src/cwt/src/main.rs
@@ -1,11 +1,15 @@
-use std::path::{Path, PathBuf};
-use std::process::{exit, Command};
+use std::path::Path;
+use std::process::exit;
 
 use buildinfo::version_string;
 use clap::Parser;
 use colored::Colorize;
 use repowalker::find_git_repo;
 use shellsetup::ShellIntegration;
+
+mod worktree;
+
+use worktree::{get_worktrees, paths_equal, Worktree};
 
 /// Exit codes for different error conditions.
 mod exit_codes {
@@ -22,9 +26,6 @@ mod exit_codes {
     /// Multiple worktrees matched the search term.
     pub const MULTIPLE_MATCHES: i32 = 6;
 }
-
-/// Length of short commit hash for display (git uses 7 by default).
-const SHORT_COMMIT_HASH_LENGTH: usize = 7;
 
 /// Macro for printing error messages that respects quiet mode.
 macro_rules! error {
@@ -122,114 +123,6 @@ struct Cli {
     quiet: bool,
 }
 
-/// Represents a single git worktree.
-#[derive(Debug, Clone)]
-struct Worktree {
-    /// The filesystem path to this worktree.
-    path: PathBuf,
-    /// The HEAD commit hash.
-    head: String,
-    /// The branch name (without refs/heads/ prefix), or None for detached HEAD.
-    branch: Option<String>,
-}
-
-impl Worktree {
-    /// Get the final directory name (e.g., "absurd-rock" from full path).
-    fn dir_name(&self) -> Option<&str> {
-        self.path.file_name()?.to_str()
-    }
-
-    /// Get the branch name for display, or short commit hash for detached HEAD.
-    fn display_branch(&self) -> String {
-        if let Some(branch) = &self.branch {
-            branch.clone()
-        } else {
-            // Show short commit hash for detached HEAD
-            let short_hash = if self.head.len() >= SHORT_COMMIT_HASH_LENGTH {
-                &self.head[..SHORT_COMMIT_HASH_LENGTH]
-            } else {
-                &self.head
-            };
-            format!("HEAD@{short_hash}")
-        }
-    }
-}
-
-/// Parses the output of `git worktree list --porcelain`.
-///
-/// The porcelain format looks like:
-/// ```text
-/// worktree /path/to/repo
-/// HEAD abc123...
-/// branch refs/heads/main
-///
-/// worktree /path/to/worktree
-/// HEAD def456...
-/// branch refs/heads/feature
-/// ```
-///
-/// For detached HEAD, the branch line is absent.
-fn parse_worktree_list(output: &str) -> Vec<Worktree> {
-    let mut worktrees = Vec::new();
-    let mut current_path: Option<PathBuf> = None;
-    let mut current_head: Option<String> = None;
-    let mut current_branch: Option<String> = None;
-
-    for line in output.lines() {
-        if line.is_empty() {
-            // End of a worktree block, save if we have the required fields.
-            // Note: .take() already leaves the Option as None, so no need to reassign.
-            if let (Some(path), Some(head)) = (current_path.take(), current_head.take()) {
-                worktrees.push(Worktree {
-                    path,
-                    head,
-                    branch: current_branch.take(),
-                });
-            }
-        } else if let Some(path) = line.strip_prefix("worktree ") {
-            current_path = Some(PathBuf::from(path));
-        } else if let Some(head) = line.strip_prefix("HEAD ") {
-            current_head = Some(head.to_string());
-        } else if let Some(branch) = line.strip_prefix("branch ") {
-            // Strip the refs/heads/ prefix
-            let branch_name = branch.strip_prefix("refs/heads/").unwrap_or(branch);
-            current_branch = Some(branch_name.to_string());
-        }
-        // Ignore other lines (like "bare" or "detached")
-    }
-
-    // Handle last block if output doesn't end with blank line
-    if let (Some(path), Some(head)) = (current_path, current_head) {
-        worktrees.push(Worktree {
-            path,
-            head,
-            branch: current_branch,
-        });
-    }
-
-    // Sort by path for consistent ordering
-    worktrees.sort_by(|a, b| a.path.cmp(&b.path));
-
-    worktrees
-}
-
-/// Gets all worktrees for the repository at the given root.
-fn get_worktrees(repo_root: &Path) -> Result<Vec<Worktree>, String> {
-    let output = Command::new("git")
-        .args(["worktree", "list", "--porcelain"])
-        .current_dir(repo_root)
-        .output()
-        .map_err(|e| format!("Failed to execute git: {e}"))?;
-
-    if !output.status.success() {
-        let stderr = String::from_utf8_lossy(&output.stderr);
-        return Err(format!("git worktree list failed: {stderr}"));
-    }
-
-    let stdout = String::from_utf8_lossy(&output.stdout);
-    Ok(parse_worktree_list(&stdout))
-}
-
 /// Finds the index of the current worktree in the sorted list.
 ///
 /// # Arguments
@@ -242,20 +135,6 @@ fn find_current_worktree(worktrees: &[Worktree], repo_root: &Path) -> Option<usi
     worktrees
         .iter()
         .position(|wt| std::fs::canonicalize(&wt.path).is_ok_and(|p| paths_equal(&p, &canonical)))
-}
-
-/// Compares two paths, handling case-insensitivity on macOS.
-fn paths_equal(a: &Path, b: &Path) -> bool {
-    // On macOS, the default filesystem is case-insensitive
-    #[cfg(target_os = "macos")]
-    {
-        a.to_string_lossy().to_lowercase() == b.to_string_lossy().to_lowercase()
-    }
-
-    #[cfg(not(target_os = "macos"))]
-    {
-        a == b
-    }
 }
 
 /// Result of searching for a worktree by name.
@@ -546,43 +425,7 @@ fn main() {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_parse_worktree_list_single() {
-        let output = "worktree /path/to/repo\nHEAD abc123\nbranch refs/heads/main\n";
-        let worktrees = parse_worktree_list(output);
-        assert_eq!(worktrees.len(), 1);
-        assert_eq!(worktrees[0].path, PathBuf::from("/path/to/repo"));
-        assert_eq!(worktrees[0].head, "abc123");
-        assert_eq!(worktrees[0].branch, Some("main".to_string()));
-    }
-
-    #[test]
-    fn test_parse_worktree_list_multiple() {
-        let output = "worktree /path/to/repo\nHEAD abc123\nbranch refs/heads/main\n\nworktree /path/to/wt\nHEAD def456\nbranch refs/heads/feature\n";
-        let worktrees = parse_worktree_list(output);
-        assert_eq!(worktrees.len(), 2);
-        assert_eq!(worktrees[0].path, PathBuf::from("/path/to/repo"));
-        assert_eq!(worktrees[1].path, PathBuf::from("/path/to/wt"));
-    }
-
-    #[test]
-    fn test_parse_worktree_list_detached_head() {
-        let output = "worktree /path/to/repo\nHEAD abc123\ndetached\n";
-        let worktrees = parse_worktree_list(output);
-        assert_eq!(worktrees.len(), 1);
-        assert_eq!(worktrees[0].branch, None);
-    }
-
-    #[test]
-    fn test_parse_worktree_list_sorted() {
-        let output = "worktree /z/repo\nHEAD abc\nbranch refs/heads/main\n\nworktree /a/repo\nHEAD def\nbranch refs/heads/feature\n";
-        let worktrees = parse_worktree_list(output);
-        assert_eq!(worktrees.len(), 2);
-        // Should be sorted by path
-        assert_eq!(worktrees[0].path, PathBuf::from("/a/repo"));
-        assert_eq!(worktrees[1].path, PathBuf::from("/z/repo"));
-    }
+    use std::path::PathBuf;
 
     #[test]
     fn test_find_worktree_by_dir_name() {
@@ -865,41 +708,6 @@ mod tests {
 
         let current = 0;
         assert_eq!(if current == 0 { count - 1 } else { current - 1 }, 2); // Wraps
-    }
-
-    #[test]
-    fn test_worktree_dir_name() {
-        let wt = Worktree {
-            path: PathBuf::from("/repo-worktrees/absurd-rock"),
-            head: "abc".to_string(),
-            branch: Some("feature".to_string()),
-        };
-        assert_eq!(wt.dir_name(), Some("absurd-rock"));
-    }
-
-    #[test]
-    fn test_worktree_display_branch() {
-        let with_branch = Worktree {
-            path: PathBuf::from("/repo"),
-            head: "abc".to_string(),
-            branch: Some("main".to_string()),
-        };
-        assert_eq!(with_branch.display_branch(), "main");
-
-        let detached = Worktree {
-            path: PathBuf::from("/repo"),
-            head: "abc1234567890".to_string(),
-            branch: None,
-        };
-        assert_eq!(detached.display_branch(), "HEAD@abc1234");
-    }
-
-    #[test]
-    fn test_parse_worktree_no_trailing_newline() {
-        let output = "worktree /path/to/repo\nHEAD abc123\nbranch refs/heads/main";
-        let worktrees = parse_worktree_list(output);
-        assert_eq!(worktrees.len(), 1);
-        assert_eq!(worktrees[0].branch, Some("main".to_string()));
     }
 
     /// Builds a worktree fixture at `path` checked out on `branch`.
